Convert SamAccountName to lowercase in salt
The hostname part in the salt must always be lowercase. Atm. it is derived from the 'sAMAccountName' attribute from LDAP, which preserves upper/lower case as used when the account was created.  Therefore, the tool does not generate the correct AES keys when used for accounts which are not completely lowercase